+++ planet-mitchell/src/ByteString/Lazy/Builder.hs
@@ -1,0 +1,5 @@
+module ByteString.Lazy.Builder
+  ( module Data.ByteString.Builder
+  ) where
+
+import Data.ByteString.Builder